Add code to decide whether the current user is an org or not
