+++ src/util/portal/portal-helper.ts
@@ -2,8 +2,15 @@ export function getPortalBuildLink(portalBaseUrl: string, appOwner: string, appN
   return `${portalBaseUrl}/users/${appOwner}/apps/${appName}/build/branches/${branchName}/builds/${buildId}`;
 }
 
-export function getPortalTestLink(portalBaseUrl: string, appOwner: string, appName: string, seriesName: string, testRunId: string): string {
-  return `${portalBaseUrl}/users/${appOwner}/apps/${appName}/test/series/${seriesName}/runs/${testRunId}`;
+export function getPortalTestLink(portalBaseUrl: string, isOrg: boolean, appOwner: string, appName: string, seriesName: string, testRunId: string): string {
+  if (isOrg)
+  {
+    return `${portalBaseUrl}/orgs/${appOwner}/apps/${appName}/test/series/${seriesName}/runs/${testRunId}`;
+  }
+  else
+  {
+    return `${portalBaseUrl}/users/${appOwner}/apps/${appName}/test/series/${seriesName}/runs/${testRunId}`;
+  }
 }
 
 export function getPortalOrgLink(portalBaseUrl: string, orgName: string): string {
